remove useless commands
